* Removed the last remenents of the OMERO rendering engine sync. * DeltaVision image sequences are "short" not "integer" values
git-svn-id: file:///home/svn/omero/trunk@1894 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/romio/src/ome/io/nio/DeltaVision.java
@@ -479,7 +479,7 @@ public class DeltaVision implements PixelBuffer {
 	 * Sets the sequence of the file. <b>Should be used for testing ONLY.</b>
 	 * @param sequence
 	 */
-	public void setSequence(int sequence)
+	public void setSequence(short sequence)
 	{
 		header.setSequence(sequence);
 	}
